requirements.txt
aerich
aiohttp
asyncpg
BeautifulSoup4
colorama
discord.py @ git+https://github.com/Rapptz/discord.py
jishaku
python-environ
PyYAML==6.0.1
tortoise-orm